requirements.txt
torch
transformers
gguf
numpy<2
pyyaml